iceberg: add partition key
Adds a wrapper around a struct_value that can be used as a partition key
in Iceberg. This partition key is constructed from a partition spec and
transformed values from a given data record.

This will be used to group records together so that parquet files can be
written partitioned by time, allowing us to remove entire parquet files
as they age out.