`rptest`: allow user to override `cloud_storage_api_endpoint`
In a CDT run with `GCP` as the cloud provider, the `cloud_storage_api_endpoint`
will be overridden by `storage.googleapis.com`, even if specified by the user.

Change the logic to allow user's choice of `cloud_storage_api_endpoint` to be
the final override value.

Default value for `SISettings.cloud_storage_api_endpoint` is still `minio-s3`,
if not otherwise specified by user or test context.