`rptest`: assert in `test_s3_oracle_self_config`
This test would previously fail in a CDT run with `GCP` as the cloud
provider.

Thanks to the changes in the previous commit, it should be passing
even in that environment, but add an `assert` as a sanity check measure.

(cherry picked from commit 65e200f31fbfe7ee3826041955222b9cf8a3a5ec)