c/tx_gateway: made end_txn fully idempotent when committing tx
When transaction is committed the process may be interrupted in any of
the steps i.e. before replicating the `preparing_commit` state update,
before requesting data partitions to commit transaction or before
replicating `completed_commit` state update. When an `end_txn` request
is retried by the client `tx_gateway` try to progress the transaction
before it is returned for further processing in either commit or abort
handling method. It may be the case that the transaction progress all
the way to `completed_commit` state, in this case Redpanda should return
a success to client as the transaction was finished.

Fixed a bug leading to the situation in which an invalid transaction
state error was returned when transaction was progressed to
`completed_commit` phase before entering a commit transaction handler.
The fix makes the `EndTxn` request fully idempotent.

Signed-off-by: Michał Maślanka <michal@redpanda.com>
(cherry picked from commit 2835bb000c56d41473add81fdf854bc81caa8e76)