s/segment_utils: pass abort source into compacted index reader
Segment compaction is a process that may take a significant amount of
time, especially when segment size is large as it was merged with other
segments during compaction. Added checking an abort source while reading
the compacted index to make the operation interruptible. This way when
requested Redpanda will be able to stop faster.

Fixes: #CORE-5542

Signed-off-by: Michał Maślanka <michal@redpanda.com>