datalake: add event multiplexer
This is a skeleton implementation of the event multiplexer. This is
a log consumer that will determine the schema ID and type (Protobuf,
JSON, Avro) of each event, and sends them to the correct translator
and writer.

In this version it translates schemaless events and sends them to a test
writer that simply counts them.