`kafka`: deal with `truncation_offset < 0` cases in `prefix_truncate()`
As a follow-up to a previous commit which failed to consider the cases
for a negative `truncation_offset`.

In Kafka, a `truncation_offset = -1` indicates a `DeleteRecords` call up to
the `high_watermark`, whereas all other negative values result in an
`offset_out_of_range` error.

Correct the cases here, and add tests for both of them.