feature_table: bumped earliest logical version
The earliest logical version is the version that the node starts with.
The earliest logical version of the previous release must be greater
than or equal to all the features that were deprecated in the future
releases. If a feature was deprecated a node assumes that it is
immediately active hence the newly formed cluster of nodes with previous
Redpanda version must bootstrap with correct earliest version.

Signed-off-by: Michał Maślanka <michal@redpanda.com>