c/tm_stm: do not sync when taking a snapshot
State machines should not wait for messages to be applied when taking a
snapshot. Waiting for application in take snapshot method may lead to a
deadlock as the apply lock is taken for the snapshot and it prevents
applying updates to state.

Another reason to remove the `sync()` call from the `take_snapshot` is
the fact that snapshoting is an optimization that happens on every node
and it is not required for correctness. Syncing state machine in
snapshot impose that the leader has to be up to date when taking a
snapshot which is not true.

Signed-off-by: Michał Maślanka <michal@redpanda.com>
(cherry picked from commit f264d14ba5db6f1a930fdd117479be811e499022)