tx/rm_stm: add best effort sequence number based fencing
if a tx is in progress, old sequence numbers are rejected until
the tx is sealed (committed/aborted). newer sequence numbers fence older
sequence numbers by aborting the trnsaction.

if a tx is already sealed, older tx_sequences are rejected and only
newer sequences are accepted.

The reason this is best effort is because sealed transaction's producer
state is prone to eviction (eg: producer limits are hit), so we are not
guaranteed to enforce this fencing all cases.