datalake: iceberg to Arrow converter
Modifies the arrow_schema_translator to also translate data.

These are combined into a single system because the schema translator
amounts to creating arrow::fields for all types. The data translator
also needs to do that, so nearly all schema translation code was
duplicated when I tried to separate them.