`admin`: wrap `cfg().to_json_single_key()` in `try/catch`
Requesting a key that does not exist in the `config_store::get()`
function throws an `std::out_of_range` exception.

This would lead to retries on an invalid request made to `get_cluster_config`
in the admin server, and what seems like a short hang to the user on the
`rpk` side.

Improve the UX by wrapping this in a `try/catch` block, and throw a
`ss::httpd::bad_param_exception` instead in case of an invalid key.

The admin request will not be retried, and the user on the `rpk` side will
immediately get a proper `400` response.