bazel: don't package shared library for libpciaccess
Because the only consumer of libpciaccess is a static library in hwloc
(which isn't bazel native, but configure_make), the shared library gets
copied to the final output, because the way in which bazel works, it
doesn't know if the static or shared library for libpciaccess was used
to build hwloc. Just don't generate the shared library as it's never
used.