tests: refactored upgrades executed in random node operations test
Change the `RandomNodeOperationsTest` upgrades handling. Previously, as
a result of membership changes during the test, it was perfectly possible
for the cluster to upgrade its cluster version while executing the test.
This might lead to a test failure as the node with previous version
wasn't able to rejoin the cluster.

Refactored the test to work with mixed versions in a way that the
majority of nodes is not upgraded throughout the test to prevent cluster
version from being updated. This way both new and old version Redpanda
nodes are able to join the cluster. Full upgrade is now executed at the
end of the test to validate if the cluster can be fully upgraded.

Signed-off-by: Michał Maślanka <michal@redpanda.com>
(cherry picked from commit 0262cdbf583ee6e55dd35a35b2725cc72429afe7)