s/batch_cache: added async_clear method to batch_cache_index
Sometimes the `batch_cache_index` may grow very large for a certain
partition segment. In this case disposing a segment abstraction may
cause a sever (up to 150ms) reactor stall. This impacts the latency and
performance of the whole system. Added method that clears the cache
asynchronously yielding if necessary.

Signed-off-by: Michał Maślanka <michal@redpanda.com>