iceberg: add non-codegen manifest_entry impl
Since the manifest entry will have a partition key embedded in it, this
introduces a manifest_entry implementation that doesn't rely on the
pre-built avrogen code.

This will also make it much easier to operate on manifest entries, as
the avrogen classes are quite limited (no comparators, etc).

This commit only introduces most of the important fields to get us
started. Most of the maps are not included, but should be in the future.