`model`: add `is_tombstone()` and modify `has_value()`
Previously, `has_value()` would consider an empty value as one with
a value, which is a confusing definition, as Kafka producers intend
for an empty value in a key-value pair to represent a tombstone.

Alter the logic of `has_value()` to consider empty values in a record
as the `false` case, and add `is_tombstone()` as a more descriptive helper
function.