tests: wait for topic to be present before starting producer
The `KgoVerifierProducer` expect topic to be present when starting. In
Kafka that is used as a source cluster in the mirror maker setup test
the topic information is propagated slowly. Added waiting for the topic
to be discoverable before proceeding with the next steps in the test.

Fixes: CORE-1639

Signed-off-by: Michał Maślanka <michal@redpanda.com>
(cherry picked from commit b4aa6fb932f54f0325d4b058621a79d85a1885fd)