sr/sharded_store: verbose messages on post schema
When an incompatible schema is posted, include the verbose compatibility
messages in the error message.

This helps users understand the reason for the incompatibility and it
matches the reference implementation.

Includes a minor fix to the way the compatibility level message is
formatted.

(cherry picked from commit cda768b4a3da734f52ab04b5220c4d93b7e5fd97)