c/members_table: include members table version in the snapshot
Members table version is updated only when cluster membership is
updated. The version must be included in the snapshot as it not always
equal to the snapshot last applied offset.

This PR makes the `members_table::version` consistent across the nodes
even if controller snapshot was applied. The version is exposed to end
users in `/v1/brokers/cluster_view` endpoint.

Fixes: CORE-6783

Signed-off-by: Michał Maślanka <michal@redpanda.com>