tests/migrations: added basic topic offload/load test
Added test validating if the topic can be released and acquired back and
the topic data are preserved. The test checks if the data that were
previously produced to the topic are available after the topic ownership was
acquired in the process of inbound migration.

Signed-off-by: Michał Maślanka <michal@redpanda.com>
(cherry picked from commit fafd06cbd601c6bd5e06ef5c3fa6a42dc67496f7)