sr/json: add detailed compatibility checks
Adapt the json compatibility check code to report the exact reason of
the incompatibility that is causing the compatibility check to fail.

Tests are adapted to now make checks explicitly on the reported
incompatibilities instead of just the boolean of whether the change is
compatible or not.

Note that because our json compatibility checks are expressed as forward
compatibility checks and not backward compatibility checks, the names of
the error types are always the inverse of what the surrounding code and
comments express. For example, we report the `additional_items_removed`
error when `newer` has `additionalItems` but `older` does not.

Some general patterns in this change are:
 * return false --> add a specific error
 * return true --> return a non-error result
 * return is_superset(...) --> depending on how the reference
   implementation works one of the following is chosen:
    - Forward the result:
        return is_superset(...)
    - Merge the result and continue to gather more:
        res.merge(is_superset(...))
    - Merge the result and add a more specific error:
        res.merge(is_superset(...));
        res.emplace(...)
    - Only add a specific error if the result has an error:
        res.emplace(...)

(cherry picked from commit 337e0671566f148e29f656606a42cd9f7540e0b7)