iceberg/tests: deflake TestSerializeManifestData
For reasons I don't fully understand (likely some flavor uninitialiezd data),
the serialized Avro bytes have a bit of non-deterministic bytes in between the
Avro header and records.

diff serialized_buf roundtrip_buf
255,256c255,256
<   00000fe0 | 22 3a 22 69 6e 74 22 7d  5d 7d 7d 5d 7d 00 23 ce  | ":"int"}]}}]}.#.
<   00000ff0 | dc 68 f6 a8 f1 4c ff 9b  6c fb 9e 11 8c 51 c8 01  | .h...L..l....Q..
---
>   00000fe0 | 22 3a 22 69 6e 74 22 7d  5d 7d 7d 5d 7d 00 2d 91  | ":"int"}]}}]}.-.
>   00000ff0 | 01 c3 65 57 53 38 21 39  28 ed cd a7 55 ab c8 01  | ..eWS8!9(...U...

Here, the ":"int"}]}}]}" is the end of the Avro header. I tried 0-initializing
the buffers in avro_iobuf_ostream, but this didn't fix the issue.

Since the issue doesn't appear to affect serialization, I'm punting on getting
byte-for-byte comparison to be stable. Instead, I've updated the test to just
check that serialization works on the roundtrip buffer.

The test would previously fail some in 1000 times with CPU stress. The updated
test passed 1000/1000 times.