iceberg: add non-codegen impl of manifest list
The avrogen::manifest_file is pretty difficult to operate on, on account
of Avro's codegen heavily using std::any and representing optional
values as unions. To manifest list operations easier (and to decouple
metadata from the Avro library) this introduces a manifest_list class
that uses simpler types to represent metadata.

This doesn't introduce any serialization code. That will come in a
subsequent commit.