rpk: add 'rpk connect' as a managed plugin
This change means that from now on, connect will
be managed by rpk, so, if a user tries to run
`rpk connect <...>` for the first time, rpk will
download, install and execute the plugin (connect).

It also introduces 3 new commands to `rpk conenct`
that live in rpk: install, uninstall, and upgrade.