iceberg: add snapshot struct
A snapshot in Iceberg represents a given table at a given point in time.
Conceptually, it is a pointer to a single manifest list, and a summary
of various metadata pertaining to that snapshot (e.g. whether it was
created by appending to the table, deleting from the table, etc).

This doesn't include code to serialize snapshots into JSON as indicated
by the spec. That will come in a later commit.