c/tx_gateway: bump producer epoch when transaction is expired
If transaction expires all the subsequent requests from the producer
with the same producer id should fail with `FENCED` error forcing
producer to bump the producer epoch and retry the request.
During the transactional coordinator refactoring this critical behavior
was missed. This way a producer wasn't able to learn about the expired
transactions.

Signed-off-by: Michał Maślanka <michal@redpanda.com>