`storage`: add `get_tombstone_delete_horizon()`
Utility function for getting the optional `timestamp` past which
tombstones can be removed.

This returns a value iff the segment `s` has been marked as cleanly
compacted, and the compaction_config has a value assigned for
`tombstone_retention_ms`. In all other cases, `std::nullopt` is returned,
indicating that tombstone records will not be removed if encountered.