ducktape: Handle incomplete batch parse
There are cases where compute_storage.py reads a batch header with some
fields > 0 (mostly CRC) while all other fields, especially the batch
size is 0. In such cases the record size is computed to -61, which when
passed to the python read function reads the entire file.

It is possible that this header is read in this way because it is being
written to while the file is parsed. There are two supporting pieces of
evidence:

* The same segment, parsed by the same script, after the test finishes,
  does not show any errors. The header where previously failure was seen
has all non zero fields now.
* The CRC matches both when the rest of the fields are 0 and in the final
version of the segment.

The change here makes the parser wait for a short period and rewind and
re-read the batch. This is done only a fixed number of times, so as to
avoid an infinite situation. This should allow the segment appender to
race ahead and write more batches for this script to consume. This
should happen until the segment is rolled.

This makes the script slower. Another option would have been to just end
the parse early in case we saw a header of batch size 0. However this
will affect test logic, this script is primarily used in a size
computation mode where we are waiting for truncation of local storage.
Under-reporting of sizes will lead us to believe that truncation
happened earlier than it actually happens.