dt: skip license nag tests in fips mode
Running the ducktape tests in FIPS mode triggers emitting the license
nag, which interferes with the tests that want to check for changes in
whether the license nag is emitted or not.

It is sufficient to run these tests in non-FIPS mode, so we can just
skip them in FIPS mode.