rpk: fix linter configuration and solve issues
This large commit reactivates the linter. In 6e3716e,
we inadvertently changed the linter to only check
issues from HEAD. However, this caused any fixes
introduced in new versions of the linter, or any
committed changes (without linting), to be ignored
by the linter.

This commit fixes that issue, pins the linter version
in our GitHub Action, and removes the golangci-lint
configuration that caused this problem in the first
place.