utils: move storage::stream_provider
I intend on moving some remote code out of cloud_storage into a new
subdirectory that doesn't depend on cloud_storage abstractions. In that
effort, I noticed that we depend on v_storage because of
storage::stream_provider.

This moves stream_provider into utils so it's easier to move stuff into
the new subdirectory.