x/sdk/js: Introduce object_builder
Helper class for building up a JavaScript object that can then be
exported as part of a module's function list.

Possible overkill for this use case (enum export), but it could
be useful in the future if we want to export a singleton object
with, for example, native function attributes.