cloud_storage: make base_manifest serialize to iobuf
I intend on making some of cloud_storage::remote more reusable. To that
end, I want to make calls to uploading the manifest use an iobuf, rather
than an input stream, as code already exists to upload iobufs.

To that end, this commit exposes the serialized manifests as an iobuf,
which all implementations already supported in building input_streams,
but didn't expose previously.