cloud_storage: upload/download_types to string
The generic code in cloud_storage::remote (e.g. upload_stream) only use
the upload_type for logging purposes. In the effort to pull the generic
code out of cloud_storage, this introduces methods to get a string_view
for these types, so that they can be passed around by the generic code
as string_views.