`cluster`: don't apply invalid property change in `topic_table`
We check that the updated topic properties post
`topic_table::apply(update_topic_properties_cmd)` is valid with the
`schema_id_validation_validator`. However, this early `co_return` with an
error code does not actually prevent the invalid configuration from being
persisted in the `topic_table`.

Rework the logic in `topic_table::apply()` to prevent persisting of an invalid
configuration.