`cluster`: add `topic_multi_property_validation()`
When applying `update_topic_properties_cmd`, there may be some checks
that depend on multiple properties. The current motivating case is
ensuring that `delete.retention.ms` is not enabled at the same time
as any tiered storage properties.

Add `topic_table::topic_multi_property_validation()` to perform these
checks and abort the topic property update if the `properties` are
found to be invalid.