iceberg: add IO wrapper for manifests
Some upcoming code will begin reading and writing manifests from remote
storage. Rather than relying directly on cloud_io, this the code in a
layer specific to Iceberg manifests.

Down the line, this will make it easier to install targeted metrics, and
abide by policies that are specific to manifests (e.g. specific timeouts
or retry configs).

This is templatized and will be reused in a subsequent commit for
manifest_lists.