cst/inv: Enable moving NTP hashes
The ntp hashes class mainly contains movable fields except for retry
chain node. Since the node is used only to set up the logger in the
object, a move-ctor is added which moves everything except for the rtc
node. Crucially, the gate of the hashes object is also moved, so we do
not stop the moved-from hashes object.