ducktape: Add inv. report parsing to scrubber test
Inventory report is added to bucket during scrubber test. This report
fudges the segment names before adding to the report for two purposes:

* at some point during the test we remove a segment and expect an
  anomaly. since we do not know the segment to be removed when
  generating the report, we have to change the name before adding the
  key, otherwise the key is found in the report and the bucket is never
  checked, thus breaking the test.
* we cannot skip adding the segment to report altogether, because if we
  do this, no hash file is generated, and as a safety measure the
  scrubber does not check for segments to avoid making many HEAD
  requests.

In the current state the addition to the test is simply a sanity check.
As more metrics are added this test can be expanded to include more
useful assertions.