`kafka`: fix `parse_and_set_shadow_indexing_mode()`
There was buggy behavior for `redpanda.remote.read` and `redpanda.remote.write`
properties and empty `std::optional<>`s created as a result of
`rpk topic alter-config [TOPIC] --delete [PROPERTY]`.

We had a bug, triggered by this sequence of events:
  1. `cloud_storage_enable_remote_read=false` at cluster level.
  2. `rpk topic create t`
  3. `rpk topic describe t` (correctly) describes `redpanda.remote.read=false`
  4. `rpk cluster config set cloud_storage_enable_remote_read=true`
  5. `rpk topic alter-config t --delete redpanda.remote.read`
  6. `rpk topic describe t` describes `redpanda.remote.read=true`, but the
      underlying configuration option is actually `std::nullopt`

We now have `rpk topic describe` printing the cluster default and indicating
topic `t` has remote read permissions, but this is not actually the case as
far as the topic is concerned (topic property is `std::nullopt`, once again,
we don't fall back on cluster default).

Fix the behavior here by removing code in `incremental_update()` that set
the shadow indexing optional to `nullopt` when the mode is
`shadow_indexing_mode::disabled`. Also, when an `alter-config --delete`
request is issued, fall back to the cluster default when setting the
new `shadow_indexing_mode`.