iceberg: initial table update class
Adds an initial table update representation, meant to be sent to the
catalog alongside a requirement, in composing atomic updates to the table.

This only introduces a subset of the updates, and only includes the
structure, no serialization or additional logic. My goal is to establish
the interface so multiple work streams can begin using these (one to
build the updates, the other to send the updates to the catalog).

Open API spec for reference:
https://github.com/apache/iceberg/blob/main/open-api/rest-catalog-open-api.yaml