iceberg: add table_requirements
Adds an initial table requirement representation, meant to be sent to
the catalog alongside a table update, in composing atomic updates to the
table.

This only introduces a subset of the requirements (those I think we'll
need for appends and table creation). My goal is to establish the
interface so multiple work streams can begin using these (one to build
the updates, the other to send the updates to the catalog).

Open API spec for reference:
https://github.com/apache/iceberg/blob/main/open-api/rest-catalog-open-api.yaml