utils: Add hybrid_clock
The hybrid_clock is a thin wrapper around the seastar::lowres_clock.
Normally it works the same way as lowres_clock except that it can be off
by 1ns compared to lowres_clock. The hybrid_clock can also act as a
manual clock if hybrdi_clock::set_manual_mode(true) was called or
the instance of manual_clock_scope is created.

The purpose of the hybrid_clock is to replace uses of the lowres_clock
and allow us to easily use manual mode in tests.

The implementation uses thread local '_manual' flag which is set to
false by default. The clock uses lowest bit of the timestamp to record
the source clock which was used to generate the timestamp. The falg is
checked if two timestamps are compared or subtracted from each other. If
both timestamps are generated using different clocks the
std::logic_error exception is thrown.

The runtime cost is minimal and prioritizes the normal mode (system
clock).