audit: Fixed audit misconfigured state logic
The audit log system requires that at the Redpanda cluster is configured
to use SASL/SCRAM authentication.  This is due to permissions applied to
the audit log topic (only permitting the audit log system to produce to
the audit log topic).  When the internal k/client attempts to connect
the the node and it reports "illegal_sasl_state", the audit system flags
that the audit system is misconfigured. This prevents audit messages
from being enqueued into the audit system and in turn ensures that the
unaudited action is not performed (that's important).

Originally, if the next error message seen is not "illegal_sasl_state"
then the flag is unset.  However, after reconnect the client may see
"broker_not_available" and then "illegal_sasl_state".  This commit
changes the behavior to wait for neither "illegal_sasl_state" nor
"broker_not_available" to be seen before unsetting the flag.

Signed-off-by: Michael Boquard <michael@redpanda.com>
(cherry picked from commit 51de7bb24d4c93e8e5aad01760f5d4de632960c6)