dt/audit: enable oauth after keycloak is up
OAuth has a background loop that polls the OIDC discovery service
(configured via the cluster config `oidc_discovery_url`) for updates.
This background loop starts as soon as OAuth is enabled via the
`sasl_mechanisms` and `http_authentication` cluster configs.
`oidc_discovery_url` has a default value pointing to a Redpanda-hosted
OIDC configuration. However, we should not rely on this Redpanda-hosted
default OIDC endpoint in our tests and CI because it might be
temporarily unavailable and cause the tests to be flaky.

To fix this test being flaky, we only enable OAuth once the local
keycloak server is up and available to serve the OIDC discovery
endpoint.