iobuf: add checks to more methods
Add the mutating method check to more mutating methods. The only check
that this does, that the "origin shard" is the same as the current
shard, also should apply to const methods, but for now I'll add it only
to mutating methods which are somehow "even worse" to call when
internal sharing is present.

Ref CORE-7061.