iceberg: table update applier
Adds a method to apply a given update to a table_metadata. There are
some basic, mechanical validations that are performed (e.g. to avoid
adding a duplicate schema).

These updates will be generated as a result of various actions (e.g.
appending to the table) and sent to the catalog. The applying of these
updates to an in-memory table_metadata allows us to validate and build
up multi-action transactions whose updates can be grouped together and
sent to the catalog together.