rpk: rename vendor to provider
vendor breaks a bunch of golang tooling because vendor is a reserved
package name for `go vendor` or vendoring third party code.

Instead of vendor, name to provider.

We only rename the package in this commit, but for clarity we will also
rename the rest of the variables and fallout from this naming in another
commit.