kafka: remove shard aggregation on consumer group metrics
Currently we only aggregate on the `shard` label in the consumer group
metrics. This aggregation doesn't result in a reduced number of metric
series. It does, however, cause seastar to use the
`metric_aggregate_by_labels` class to aggregate the metrics.

This class uses a contiguous map type that isn't easily replaced and in
cases where there are many consumer groups, kafka topics, and partitions
there is enough unique label sets to cause an oversized allocation in
the continguous map.

By removing the `shard` label from aggregation though we avoid the code
path where `metric_aggregate_by_labels` is used and hence avoid the
oversized allocation.