http: Introduce a rest-entity helper
The rest-entity represents a REST resource (a path/url) and exposes a
set of CRUD operations on it. This is a pure virtual class, and classes
extending from it must provide the resource name which will be used to
create methods and helpers.

For example given a resource rest_entity{"person"} which has the
resource name "person" and the root url "/api/v1", the following
convenience methods are available:

GET /api/v1/person
POST /api/v1/person
DELETE /api/v1/person/foo
GET /api/v1/person/bar

The http verbs can be overridden and optional header and query params
can be supplied.