http: Make http-client abstract
The http client now extends a pure virtual class. This virtual
class/interface exposes a single method which fully reads a response
before returning it as part of an aggregate struct. This makes testing
and mocking the http client easier specifically for requests with short
responses. Intended to be used with the iceberg rest-client primarily.