features: new metric enterprise_license_expiry_sec
Adds a metric to allow easier monitoring of when the enterprise license
is going to expiry.

The dependency on the `security` module existed even previously in the
bazel build but it was missing from the cmake build.

(cherry picked from commit edb5f48493aaec25ebcb10f35eb26b3ad5ae1a10)