cloud-tests: Fix incomplete kubectl command
We were not passing a resource type to get (`pod` missing) and hence the
command would fail on operator v1. This does however work on operator v2
where a redpanda CRD exists.

Fix this by just checking whether the command fails or not and replace
some usages with the existing helper function.