dt: install enterprise license for upgrade tests
The commits in this PR change Redpanda to require an enterprise license
on upgrade, which means that ducktape tests that use enterprise features
and are upgrading nodes also require a license.

Therefore, this installs a license in all existing tests that require a
license. Note that we don't install a license in all tests, but only in
those that require one. This is to ensure that during local development
we can conveniently run any test without a license that doesn't require
one.