iceberg: manifest file packing
Iceberg has two modes for appending data:
- Fast append: new files are always added to a brand new manifest, and
  the new manifest is added to the manifest list.
- Merge append: manifests are bin-packed to a target size (e.g. 8MiB),
  and bins of manifests are merged to create larger manifests, thereby
  reducing the number of manifests in a given snapshot.

While fast appends are easier to implement, they are not scalable for a
streaming system like Redpanda, especially while we don't have
compaction. To limit the rate of metadata generated, Redpanda will merge
append by default.

To that end, this introduces a simple bin packer for manifest files to
group together manifests to a target size. The implementation is a
simplified version of a more generic version in the Icberg python
library (functionally it is the same though).