config: add a cluster taint for experimental features
Adds the configuration option:

  enable_experimental_unrecoverable_data_corrupting_features

which is intended to be used during development, only, and never in a
production cluster. When enabled this property does not turn on any
specific experimental feature, but rather is used to allow the
enablement of configuration properties associated with experimental
features.

Despite the name, we take extra precautions to avoid accidentally
enabling this feature (e.g. copy-paste configuration changes). While the
configuration option has boolean semantics, in order to enable this
property it must be set to a special value:

  the current unix epoch expressed in seconds, with a +/- 1 hour
  tolerance. the time is compared to the current time on the broker.

In order to support in-tree development of features that span releases,
we need a mechanism for controlling access to features that should not
be enabled in production clusters.

After it is enabled, this property cannot be disabled, acting like a
taint on the cluster.

This flag vs an environment variable? An environment variable that
enables developer features is also a viable option. The downside to this
is that it may be a bit painful to enable this in a cloud environment
when testing as changing the environment variables may mean changing k8s
pod specs.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>